Switch out of locked courses feature
CoursesController.php edited to improve the actual enrollment numbers. Unserializing the f_schedule array is problematic for students registered in more than one IDH 3034 (i.e. students taking summer AND fall classes). So this simply counts the occurance of the substring representing the course in all the f_schedules. It's actually faster and more accurate. A similar approach is used to query the registree students.

admin_registrees.ctp edited to show if a student was locked in the class